feat(web): extract submission spec helpers into submission-spec-lib
Move the category field matrix, preflight checks, and packet builder into
a dedicated -lib module with expanded Vitest coverage while preserving the
public @/lib/submission-spec import surface.

Closes #4288
